Working Class
This class counts the points that are in the circle, outside the
circle, and the total points overall. On ever 100th iteration, the area
is calculated and the number of points in the circle, outside the
circle, and total up to that point are printed. To get the X and the Y
coordinates, I generate a random double.